+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [4.2.1 (unreleased)](https://github.com/kdeldycke/workflows/compare/v4.2.0...main)
+
+> \[!IMPORTANT\]
+> This version is not released yet and is under active development.
+
 ## [4.2.0 (2024-07-15)](https://github.com/kdeldycke/workflows/compare/v4.1.4...v4.2.0)
 
 - Rename `gha-utils mailmap` command to `gha-utils mailmap-sync`.